Move purpose-process tests to integration dir
