+++ packages/purpose-process/package.json
@@ -5,7 +5,7 @@
   "main": "dist",
   "type": "module",
   "scripts": {
-    "test": "vitest",
+    "test": "vitest --config vitest.integration.config.ts",
     "lint": "eslint . --ext .ts,.tsx",
     "lint:autofix": "eslint . --ext .ts,.tsx --fix",
     "format:check": "prettier --check src",
