+++ packages/purpose-process/test/integration/archivePurposeVersion.test.ts
@@ -1,101 +1,75 @@
 /* eslint-disable @typescript-eslint/no-floating-promises */
-import { describe, expect, it } from "vitest";
 import {
-  Agreement,
-  DelegationId,
-  DraftPurposeDeletedV2,
-  EService,
-  Purpose,
-  PurposeId,
-  PurposeVersion,
-  TenantId,
-  WaitingForApprovalPurposeDeletedV2,
-  delegationKind,
-  delegationState,
-  generateId,
-  purposeVersionState,
-  tenantKind,
-  toPurposeV2,
-  toReadModelEService,
-} from "pagopa-interop-models";
-import {
-  getMockPurpose,
-  writeInReadmodel,
-  decodeProtobufPayload,
   getMockPurposeVersion,
+  getMockPurpose,
+  decodeProtobufPayload,
   getMockAuthData,
   getMockDelegation,
   addSomeRandomDelegations,
-  getMockTenant,
   getMockAgreement,
+  getMockEService,
+  getMockTenant,
   getMockContext,
 } from "pagopa-interop-commons-test";
 import {
+  PurposeVersion,
+  purposeVersionState,
+  Purpose,
+  generateId,
+  PurposeArchivedV2,
+  toPurposeV2,
+  PurposeId,
+  PurposeVersionId,
+  TenantId,
+  toPurposeVersionV2,
+  delegationKind,
+  delegationState,
+  Agreement,
+  EService,
+  eserviceMode,
+  tenantKind,
+  agreementState,
+  DelegationId,
+} from "pagopa-interop-models";
+import { describe, expect, it, vi } from "vitest";
+import {
   purposeNotFound,
   organizationIsNotTheConsumer,
-  purposeCannotBeDeleted,
+  purposeVersionNotFound,
+  notValidVersionState,
   organizationIsNotTheDelegatedConsumer,
   puroposeDelegationNotFound,
-} from "../src/model/domain/errors.js";
+} from "../../src/model/domain/errors.js";
 import {
   addOneAgreement,
   addOneDelegation,
   addOneEService,
   addOnePurpose,
   addOneTenant,
-  eservices,
-  getMockEService,
   purposeService,
   readLastPurposeEvent,
-} from "./utils.js";
+} from "../integrationUtils.js";
 
-describe("deletePurpose", () => {
-  it("should write on event-store for the deletion of a purpose (no versions)", async () => {
-    const mockEService = getMockEService();
-    const mockPurpose: Purpose = {
-      ...getMockPurpose(),
-      eserviceId: mockEService.id,
-      versions: [],
+describe("archivePurposeVersion", () => {
+  it("should write on event-store for the archiving of a purpose version", async () => {
+    vi.useFakeTimers();
+    vi.setSystemTime(new Date());
+
+    const mockPurposeVersion: PurposeVersion = {
+      ...getMockPurposeVersion(),
+      state: purposeVersionState.active,
     };
-
-    await addOnePurpose(mockPurpose);
-    await writeInReadmodel(toReadModelEService(mockEService), eservices);
-
-    await purposeService.deletePurpose(
-      mockPurpose.id,
-      getMockContext({ authData: getMockAuthData(mockPurpose.consumerId) })
-    );
-
-    const writtenEvent = await readLastPurposeEvent(mockPurpose.id);
-
-    expect(writtenEvent).toMatchObject({
-      stream_id: mockPurpose.id,
-      version: "1",
-      type: "DraftPurposeDeleted",
-      event_version: 2,
-    });
-
-    const writtenPayload = decodeProtobufPayload({
-      messageType: DraftPurposeDeletedV2,
-      payload: writtenEvent.data,
-    });
-
-    expect(writtenPayload.purpose).toEqual(toPurposeV2(mockPurpose));
-  });
-  it("should write on event-store for the deletion of a purpose (draft version)", async () => {
-    const mockEService = getMockEService();
-    const mockPurposeVersion = getMockPurposeVersion(purposeVersionState.draft);
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
-      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
     };
-
     await addOnePurpose(mockPurpose);
-    await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
-    await purposeService.deletePurpose(
-      mockPurpose.id,
+    const returnedPurposeVersion = await purposeService.archivePurposeVersion(
+      {
+        purposeId: mockPurpose.id,
+        versionId: mockPurposeVersion.id,
+      },
       getMockContext({ authData: getMockAuthData(mockPurpose.consumerId) })
     );
 
@@ -104,33 +78,59 @@ describe("deletePurpose", () => {
     expect(writtenEvent).toMatchObject({
       stream_id: mockPurpose.id,
       version: "1",
-      type: "DraftPurposeDeleted",
+      type: "PurposeArchived",
       event_version: 2,
     });
 
-    const writtenPayload = decodeProtobufPayload({
-      messageType: DraftPurposeDeletedV2,
-      payload: writtenEvent.data,
-    });
-
-    expect(writtenPayload.purpose).toEqual(toPurposeV2(mockPurpose));
-  });
-  it("should write on event-store for the deletion of a purpose (waitingForApproval version)", async () => {
-    const mockEService = getMockEService();
-    const mockPurposeVersion = getMockPurposeVersion(
-      purposeVersionState.waitingForApproval
-    );
-    const mockPurpose: Purpose = {
-      ...getMockPurpose(),
-      eserviceId: mockEService.id,
-      versions: [mockPurposeVersion],
+    const expectedPurpose: Purpose = {
+      ...mockPurpose,
+      versions: [
+        {
+          ...mockPurposeVersion,
+          state: purposeVersionState.archived,
+          updatedAt: new Date(),
+        },
+      ],
+      updatedAt: new Date(),
     };
 
-    await addOnePurpose(mockPurpose);
-    await writeInReadmodel(toReadModelEService(mockEService), eservices);
+    const writtenPayload = decodeProtobufPayload({
+      messageType: PurposeArchivedV2,
+      payload: writtenEvent.data,
+    });
 
-    await purposeService.deletePurpose(
-      mockPurpose.id,
+    expect(writtenPayload.purpose).toEqual(toPurposeV2(expectedPurpose));
+    expect(
+      writtenPayload.purpose?.versions.find(
+        (v) => v.id === returnedPurposeVersion.id
+      )
+    ).toEqual(toPurposeVersionV2(returnedPurposeVersion));
+
+    vi.useRealTimers();
+  });
+  it("should write on event-store for the archiving of a purpose version, and delete waitingForApproval versions", async () => {
+    vi.useFakeTimers();
+    vi.setSystemTime(new Date());
+
+    const mockPurposeVersion1: PurposeVersion = {
+      ...getMockPurposeVersion(),
+      state: purposeVersionState.active,
+    };
+    const mockPurposeVersion2: PurposeVersion = {
+      ...getMockPurposeVersion(),
+      state: purposeVersionState.waitingForApproval,
+    };
+    const mockPurpose: Purpose = {
+      ...getMockPurpose(),
+      versions: [mockPurposeVersion1, mockPurposeVersion2],
+    };
+    await addOnePurpose(mockPurpose);
+
+    const returnedPurposeVersion = await purposeService.archivePurposeVersion(
+      {
+        purposeId: mockPurpose.id,
+        versionId: mockPurposeVersion1.id,
+      },
       getMockContext({ authData: getMockAuthData(mockPurpose.consumerId) })
     );
 
@@ -139,27 +139,51 @@ describe("deletePurpose", () => {
     expect(writtenEvent).toMatchObject({
       stream_id: mockPurpose.id,
       version: "1",
-      type: "WaitingForApprovalPurposeDeleted",
+      type: "PurposeArchived",
       event_version: 2,
     });
 
+    const expectedPurpose: Purpose = {
+      ...mockPurpose,
+      versions: [
+        {
+          ...mockPurposeVersion1,
+          state: purposeVersionState.archived,
+          updatedAt: new Date(),
+        },
+      ],
+      updatedAt: new Date(),
+    };
+
     const writtenPayload = decodeProtobufPayload({
-      messageType: WaitingForApprovalPurposeDeletedV2,
+      messageType: PurposeArchivedV2,
       payload: writtenEvent.data,
     });
 
-    expect(writtenPayload.purpose).toEqual(toPurposeV2(mockPurpose));
+    expect(writtenPayload.purpose).toEqual(toPurposeV2(expectedPurpose));
+    expect(
+      writtenPayload.purpose?.versions.find(
+        (v) => v.id === returnedPurposeVersion.id
+      )
+    ).toEqual(toPurposeVersionV2(returnedPurposeVersion));
+
+    vi.useRealTimers();
   });
-  it("should succeed when requester is Consumer Delegate and the Purpose is in a deletable state", async () => {
-    const authData = getMockAuthData();
-    const mockEService = getMockEService();
-    const mockPurposeVersion = getMockPurposeVersion(purposeVersionState.draft);
+  it("should write on event-store for the archiving of a purpose version when requester is Consumer Delegate", async () => {
+    vi.useFakeTimers();
+    vi.setSystemTime(new Date());
+
+    const mockPurposeVersion: PurposeVersion = {
+      ...getMockPurposeVersion(),
+      state: purposeVersionState.active,
+    };
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
-      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
       delegationId: generateId<DelegationId>(),
     };
+
+    const authData = getMockAuthData();
 
     const delegation = getMockDelegation({
       id: mockPurpose.delegationId,
@@ -173,10 +197,12 @@ describe("deletePurpose", () => {
     await addOnePurpose(mockPurpose);
     await addOneDelegation(delegation);
     await addSomeRandomDelegations(mockPurpose, addOneDelegation);
-    await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
-    await purposeService.deletePurpose(
-      mockPurpose.id,
+    const returnedPurposeVersion = await purposeService.archivePurposeVersion(
+      {
+        purposeId: mockPurpose.id,
+        versionId: mockPurposeVersion.id,
+      },
       getMockContext({ authData })
     );
 
@@ -185,18 +211,40 @@ describe("deletePurpose", () => {
     expect(writtenEvent).toMatchObject({
       stream_id: mockPurpose.id,
       version: "1",
-      type: "DraftPurposeDeleted",
+      type: "PurposeArchived",
       event_version: 2,
     });
 
+    const expectedPurpose: Purpose = {
+      ...mockPurpose,
+      versions: [
+        {
+          ...mockPurposeVersion,
+          state: purposeVersionState.archived,
+          updatedAt: new Date(),
+        },
+      ],
+      updatedAt: new Date(),
+    };
+
     const writtenPayload = decodeProtobufPayload({
-      messageType: DraftPurposeDeletedV2,
+      messageType: PurposeArchivedV2,
       payload: writtenEvent.data,
     });
 
-    expect(writtenPayload.purpose).toEqual(toPurposeV2(mockPurpose));
+    expect(writtenPayload.purpose).toEqual(toPurposeV2(expectedPurpose));
+    expect(
+      writtenPayload.purpose?.versions.find(
+        (v) => v.id === returnedPurposeVersion.id
+      )
+    ).toEqual(toPurposeVersionV2(returnedPurposeVersion));
+
+    vi.useRealTimers();
   });
-  it("should succeed when requester is Consumer Delegate and the eservice was created by a delegated tenant and the Purpose is in a deletable state", async () => {
+  it("should write on event-store for the archiving of a purpose version when requester is Consumer Delegate and the eservice was created by a delegated tenant", async () => {
+    vi.useFakeTimers();
+    vi.setSystemTime(new Date());
+
     const producer = {
       ...getMockTenant(),
       id: generateId<TenantId>(),
@@ -220,6 +268,7 @@ describe("deletePurpose", () => {
 
     const eservice: EService = {
       ...getMockEService(),
+      mode: eserviceMode.receive,
       producerId: producer.id,
     };
     const agreement: Agreement = {
@@ -227,9 +276,13 @@ describe("deletePurpose", () => {
       producerId: producer.id,
       consumerId: consumer.id,
       eserviceId: eservice.id,
+      state: agreementState.active,
     };
 
-    const mockPurposeVersion = getMockPurposeVersion(purposeVersionState.draft);
+    const mockPurposeVersion: PurposeVersion = {
+      ...getMockPurposeVersion(),
+      state: purposeVersionState.active,
+    };
 
     const delegatePurpose: Purpose = {
       ...getMockPurpose(),
@@ -267,8 +320,11 @@ describe("deletePurpose", () => {
     await addOneDelegation(consumerDelegation);
     await addSomeRandomDelegations(delegatePurpose, addOneDelegation);
 
-    await purposeService.deletePurpose(
-      delegatePurpose.id,
+    const returnedPurposeVersion = await purposeService.archivePurposeVersion(
+      {
+        purposeId: delegatePurpose.id,
+        versionId: mockPurposeVersion.id,
+      },
       getMockContext({ authData: getMockAuthData(consumerDelegate.id) })
     );
 
@@ -277,92 +333,139 @@ describe("deletePurpose", () => {
     expect(writtenEvent).toMatchObject({
       stream_id: delegatePurpose.id,
       version: "1",
-      type: "DraftPurposeDeleted",
+      type: "PurposeArchived",
       event_version: 2,
     });
 
+    const expectedPurpose: Purpose = {
+      ...delegatePurpose,
+      versions: [
+        {
+          ...mockPurposeVersion,
+          state: purposeVersionState.archived,
+          updatedAt: new Date(),
+        },
+      ],
+      updatedAt: new Date(),
+    };
+
     const writtenPayload = decodeProtobufPayload({
-      messageType: DraftPurposeDeletedV2,
+      messageType: PurposeArchivedV2,
       payload: writtenEvent.data,
     });
 
-    expect(writtenPayload.purpose).toEqual(toPurposeV2(delegatePurpose));
-  });
-  it("should throw purposeNotFound if the purpose doesn't exist", async () => {
-    const randomId: PurposeId = generateId();
-    const mockPurpose = getMockPurpose();
-
-    await addOnePurpose(mockPurpose);
+    expect(writtenPayload.purpose).toEqual(toPurposeV2(expectedPurpose));
     expect(
-      purposeService.deletePurpose(
-        randomId,
+      writtenPayload.purpose?.versions.find(
+        (v) => v.id === returnedPurposeVersion.id
+      )
+    ).toEqual(toPurposeVersionV2(returnedPurposeVersion));
+
+    vi.useRealTimers();
+  });
+
+  it("should throw purposeNotFound if the purpose doesn't exist", async () => {
+    const randomPurposeId: PurposeId = generateId();
+    const randomVersionId: PurposeVersionId = generateId();
+    const mockPurpose = getMockPurpose();
+    await addOnePurpose(mockPurpose);
+
+    expect(
+      purposeService.archivePurposeVersion(
+        {
+          purposeId: randomPurposeId,
+          versionId: randomVersionId,
+        },
         getMockContext({ authData: getMockAuthData(mockPurpose.consumerId) })
       )
-    ).rejects.toThrowError(purposeNotFound(randomId));
+    ).rejects.toThrowError(purposeNotFound(randomPurposeId));
   });
   it("should throw organizationIsNotTheConsumer if the requester is not the consumer", async () => {
-    const mockEService = getMockEService();
-    const mockPurposeVersion: PurposeVersion = getMockPurposeVersion(
-      purposeVersionState.draft
-    );
+    const randomOrganizationId: TenantId = generateId();
+
+    const mockPurposeVersion: PurposeVersion = {
+      ...getMockPurposeVersion(),
+      state: purposeVersionState.active,
+    };
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
-      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
     };
 
     await addOnePurpose(mockPurpose);
-    await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
     expect(
-      purposeService.deletePurpose(
-        mockPurpose.id,
-        getMockContext({ authData: getMockAuthData(mockEService.producerId) })
+      purposeService.archivePurposeVersion(
+        {
+          purposeId: mockPurpose.id,
+          versionId: mockPurposeVersion.id,
+        },
+        getMockContext({ authData: getMockAuthData(randomOrganizationId) })
+      )
+    ).rejects.toThrowError(organizationIsNotTheConsumer(randomOrganizationId));
+  });
+  it("should throw purposeVersionNotFound if the purpose version doesn't exist", async () => {
+    const randomVersionId: PurposeVersionId = generateId();
+    const mockPurpose: Purpose = {
+      ...getMockPurpose(),
+      versions: [],
+    };
+
+    await addOnePurpose(mockPurpose);
+
+    expect(
+      purposeService.archivePurposeVersion(
+        {
+          purposeId: mockPurpose.id,
+          versionId: randomVersionId,
+        },
+        getMockContext({ authData: getMockAuthData(mockPurpose.consumerId) })
       )
     ).rejects.toThrowError(
-      organizationIsNotTheConsumer(mockEService.producerId)
+      purposeVersionNotFound(mockPurpose.id, randomVersionId)
     );
   });
   it.each(
     Object.values(purposeVersionState).filter(
       (state) =>
-        state !== purposeVersionState.waitingForApproval &&
-        state !== purposeVersionState.draft
+        state !== purposeVersionState.active &&
+        state !== purposeVersionState.suspended
     )
   )(
-    "should throw purposeCannotBeDeleted if the purpose has a $s version ",
+    "should throw notValidVersionState if the purpose version is in %s state",
     async (state) => {
-      const mockEService = getMockEService();
       const mockPurposeVersion = getMockPurposeVersion(state);
 
       const mockPurpose: Purpose = {
         ...getMockPurpose(),
-        eserviceId: mockEService.id,
         versions: [mockPurposeVersion],
       };
 
       await addOnePurpose(mockPurpose);
-      await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
       expect(
-        purposeService.deletePurpose(
-          mockPurpose.id,
+        purposeService.archivePurposeVersion(
+          {
+            purposeId: mockPurpose.id,
+            versionId: mockPurposeVersion.id,
+          },
           getMockContext({
             authData: getMockAuthData(mockPurpose.consumerId),
           })
         )
-      ).rejects.toThrowError(purposeCannotBeDeleted(mockPurpose.id));
+      ).rejects.toThrowError(
+        notValidVersionState(mockPurposeVersion.id, mockPurposeVersion.state)
+      );
     }
   );
-  it("should throw organizationIsNotTheDelegatedConsumer when the requester is the Consumer and is deleting a purpose created by the delegate in deletePurpose", async () => {
+  it("should throw organizationIsNotTheDelegatedConsumer when the requester is the Consumer and is archiving a purpose version created by the delegate in archivePurposeVersion", async () => {
     const authData = getMockAuthData();
-    const mockEService = getMockEService();
-    const mockPurposeVersion: PurposeVersion = getMockPurposeVersion(
-      purposeVersionState.draft
-    );
+    const mockPurposeVersion: PurposeVersion = {
+      ...getMockPurposeVersion(),
+      state: purposeVersionState.active,
+    };
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
-      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
       consumerId: authData.organizationId,
       delegationId: generateId<DelegationId>(),
@@ -378,19 +481,23 @@ describe("deletePurpose", () => {
     });
     await addOnePurpose(mockPurpose);
     await addOneDelegation(delegation);
-    await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
     expect(
-      purposeService.deletePurpose(mockPurpose.id, getMockContext({ authData }))
+      purposeService.archivePurposeVersion(
+        {
+          purposeId: mockPurpose.id,
+          versionId: mockPurposeVersion.id,
+        },
+        getMockContext({ authData })
+      )
     ).rejects.toThrowError(
       organizationIsNotTheDelegatedConsumer(
         authData.organizationId,
-        delegation.id
+        mockPurpose.delegationId
       )
     );
   });
-
-  it("should throw puroposeDelegationNotFound when the requester is the Consumer, is deleting a purpose created by a delegate in deletePurpose, but the delegation cannot be found", async () => {
+  it("should throw puroposeDelegationNotFound when the requester is the Consumer, is archiving a purpose created by a delegate, but the delegation cannot be found", async () => {
     const authData = getMockAuthData();
     const mockEService = getMockEService();
     const mockPurposeVersion: PurposeVersion = getMockPurposeVersion(
@@ -405,27 +512,31 @@ describe("deletePurpose", () => {
     };
 
     await addOnePurpose(mockPurpose);
-    await writeInReadmodel(toReadModelEService(mockEService), eservices);
-
+    await addOneEService(mockEService);
     expect(
-      purposeService.deletePurpose(
-        mockPurpose.id,
-        getMockContext({ authData: getMockAuthData(mockEService.producerId) })
+      purposeService.archivePurposeVersion(
+        {
+          purposeId: mockPurpose.id,
+          versionId: mockPurposeVersion.id,
+        },
+        getMockContext({ authData })
       )
     ).rejects.toThrowError(
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       puroposeDelegationNotFound(mockPurpose.id, mockPurpose.delegationId!)
     );
   });
-  it("should throw organizationIsNotTheConsumer if the requester is a delegate for the eservice and there is no delegationId in the purpose", async () => {
-    const mockEService = getMockEService();
-    const mockPurposeVersion: PurposeVersion = getMockPurposeVersion(
-      purposeVersionState.draft
-    );
+
+  it("should throw organizationIsNotTheConsumer when the requester is a delegate for the eservice and there is no delegationId in the purpose in archivePurposeVersion", async () => {
+    const authData = getMockAuthData();
+    const mockPurposeVersion: PurposeVersion = {
+      ...getMockPurposeVersion(),
+      state: purposeVersionState.active,
+    };
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
-      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
+      consumerId: authData.organizationId,
       delegationId: undefined,
     };
 
@@ -438,25 +549,29 @@ describe("deletePurpose", () => {
     });
     await addOnePurpose(mockPurpose);
     await addOneDelegation(delegation);
-    await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
     expect(
-      purposeService.deletePurpose(
-        mockPurpose.id,
+      purposeService.archivePurposeVersion(
+        {
+          purposeId: mockPurpose.id,
+          versionId: mockPurposeVersion.id,
+        },
         getMockContext({ authData: getMockAuthData(delegation.delegateId) })
       )
     ).rejects.toThrowError(organizationIsNotTheConsumer(delegation.delegateId));
   });
 
   it("should throw organizationIsNotTheDelegatedConsumer if the the requester is a delegate for the eservice and there is a delegationId in purpose but for a different delegationId (a different delegate)", async () => {
+    const authData = getMockAuthData();
     const mockEService = getMockEService();
-    const mockPurposeVersion: PurposeVersion = getMockPurposeVersion(
-      purposeVersionState.draft
-    );
+    const mockPurposeVersion: PurposeVersion = {
+      ...getMockPurposeVersion(),
+      state: purposeVersionState.active,
+    };
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
-      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
+      consumerId: authData.organizationId,
       delegationId: generateId<DelegationId>(),
     };
 
@@ -481,11 +596,14 @@ describe("deletePurpose", () => {
     await addOnePurpose(mockPurpose);
     await addOneDelegation(delegation);
     await addOneDelegation(purposeDelegation);
-    await writeInReadmodel(toReadModelEService(mockEService), eservices);
+    await addOneEService(mockEService);
 
     expect(
-      purposeService.deletePurpose(
-        mockPurpose.id,
+      purposeService.archivePurposeVersion(
+        {
+          purposeId: mockPurpose.id,
+          versionId: mockPurposeVersion.id,
+        },
         getMockContext({ authData: getMockAuthData(delegation.delegateId) })
       )
     ).rejects.toThrowError(
