+++ packages/purpose-process/test/integration/deletePurpose.test.ts
@@ -1,75 +1,101 @@
 /* eslint-disable @typescript-eslint/no-floating-promises */
+import { describe, expect, it } from "vitest";
 import {
-  getMockPurposeVersion,
+  Agreement,
+  DelegationId,
+  DraftPurposeDeletedV2,
+  EService,
+  Purpose,
+  PurposeId,
+  PurposeVersion,
+  TenantId,
+  WaitingForApprovalPurposeDeletedV2,
+  delegationKind,
+  delegationState,
+  generateId,
+  purposeVersionState,
+  tenantKind,
+  toPurposeV2,
+  toReadModelEService,
+} from "pagopa-interop-models";
+import {
   getMockPurpose,
+  writeInReadmodel,
   decodeProtobufPayload,
+  getMockPurposeVersion,
   getMockAuthData,
   getMockDelegation,
   addSomeRandomDelegations,
-  getMockAgreement,
-  getMockEService,
   getMockTenant,
+  getMockAgreement,
   getMockContext,
 } from "pagopa-interop-commons-test";
 import {
-  PurposeVersion,
-  purposeVersionState,
-  Purpose,
-  generateId,
-  PurposeArchivedV2,
-  toPurposeV2,
-  PurposeId,
-  PurposeVersionId,
-  TenantId,
-  toPurposeVersionV2,
-  delegationKind,
-  delegationState,
-  Agreement,
-  EService,
-  eserviceMode,
-  tenantKind,
-  agreementState,
-  DelegationId,
-} from "pagopa-interop-models";
-import { describe, expect, it, vi } from "vitest";
-import {
   purposeNotFound,
   organizationIsNotTheConsumer,
-  purposeVersionNotFound,
-  notValidVersionState,
+  purposeCannotBeDeleted,
   organizationIsNotTheDelegatedConsumer,
   puroposeDelegationNotFound,
-} from "../src/model/domain/errors.js";
+} from "../../src/model/domain/errors.js";
 import {
   addOneAgreement,
   addOneDelegation,
   addOneEService,
   addOnePurpose,
   addOneTenant,
+  eservices,
   purposeService,
   readLastPurposeEvent,
-} from "./utils.js";
+} from "../integrationUtils.js";
+import { getMockEService } from "../mockUtils.js";
 
-describe("archivePurposeVersion", () => {
-  it("should write on event-store for the archiving of a purpose version", async () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date());
-
-    const mockPurposeVersion: PurposeVersion = {
-      ...getMockPurposeVersion(),
-      state: purposeVersionState.active,
-    };
+describe("deletePurpose", () => {
+  it("should write on event-store for the deletion of a purpose (no versions)", async () => {
+    const mockEService = getMockEService();
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
+      eserviceId: mockEService.id,
+      versions: [],
+    };
+
+    await addOnePurpose(mockPurpose);
+    await writeInReadmodel(toReadModelEService(mockEService), eservices);
+
+    await purposeService.deletePurpose(
+      mockPurpose.id,
+      getMockContext({ authData: getMockAuthData(mockPurpose.consumerId) })
+    );
+
+    const writtenEvent = await readLastPurposeEvent(mockPurpose.id);
+
+    expect(writtenEvent).toMatchObject({
+      stream_id: mockPurpose.id,
+      version: "1",
+      type: "DraftPurposeDeleted",
+      event_version: 2,
+    });
+
+    const writtenPayload = decodeProtobufPayload({
+      messageType: DraftPurposeDeletedV2,
+      payload: writtenEvent.data,
+    });
+
+    expect(writtenPayload.purpose).toEqual(toPurposeV2(mockPurpose));
+  });
+  it("should write on event-store for the deletion of a purpose (draft version)", async () => {
+    const mockEService = getMockEService();
+    const mockPurposeVersion = getMockPurposeVersion(purposeVersionState.draft);
+    const mockPurpose: Purpose = {
+      ...getMockPurpose(),
+      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
     };
-    await addOnePurpose(mockPurpose);
 
-    const returnedPurposeVersion = await purposeService.archivePurposeVersion(
-      {
-        purposeId: mockPurpose.id,
-        versionId: mockPurposeVersion.id,
-      },
+    await addOnePurpose(mockPurpose);
+    await writeInReadmodel(toReadModelEService(mockEService), eservices);
+
+    await purposeService.deletePurpose(
+      mockPurpose.id,
       getMockContext({ authData: getMockAuthData(mockPurpose.consumerId) })
     );
 
@@ -78,59 +104,33 @@ describe("archivePurposeVersion", () => {
     expect(writtenEvent).toMatchObject({
       stream_id: mockPurpose.id,
       version: "1",
-      type: "PurposeArchived",
+      type: "DraftPurposeDeleted",
       event_version: 2,
     });
 
-    const expectedPurpose: Purpose = {
-      ...mockPurpose,
-      versions: [
-        {
-          ...mockPurposeVersion,
-          state: purposeVersionState.archived,
-          updatedAt: new Date(),
-        },
-      ],
-      updatedAt: new Date(),
-    };
-
     const writtenPayload = decodeProtobufPayload({
-      messageType: PurposeArchivedV2,
+      messageType: DraftPurposeDeletedV2,
       payload: writtenEvent.data,
     });
 
-    expect(writtenPayload.purpose).toEqual(toPurposeV2(expectedPurpose));
-    expect(
-      writtenPayload.purpose?.versions.find(
-        (v) => v.id === returnedPurposeVersion.id
-      )
-    ).toEqual(toPurposeVersionV2(returnedPurposeVersion));
-
-    vi.useRealTimers();
+    expect(writtenPayload.purpose).toEqual(toPurposeV2(mockPurpose));
   });
-  it("should write on event-store for the archiving of a purpose version, and delete waitingForApproval versions", async () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date());
-
-    const mockPurposeVersion1: PurposeVersion = {
-      ...getMockPurposeVersion(),
-      state: purposeVersionState.active,
-    };
-    const mockPurposeVersion2: PurposeVersion = {
-      ...getMockPurposeVersion(),
-      state: purposeVersionState.waitingForApproval,
-    };
+  it("should write on event-store for the deletion of a purpose (waitingForApproval version)", async () => {
+    const mockEService = getMockEService();
+    const mockPurposeVersion = getMockPurposeVersion(
+      purposeVersionState.waitingForApproval
+    );
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
-      versions: [mockPurposeVersion1, mockPurposeVersion2],
+      eserviceId: mockEService.id,
+      versions: [mockPurposeVersion],
     };
-    await addOnePurpose(mockPurpose);
 
-    const returnedPurposeVersion = await purposeService.archivePurposeVersion(
-      {
-        purposeId: mockPurpose.id,
-        versionId: mockPurposeVersion1.id,
-      },
+    await addOnePurpose(mockPurpose);
+    await writeInReadmodel(toReadModelEService(mockEService), eservices);
+
+    await purposeService.deletePurpose(
+      mockPurpose.id,
       getMockContext({ authData: getMockAuthData(mockPurpose.consumerId) })
     );
 
@@ -139,51 +139,27 @@ describe("archivePurposeVersion", () => {
     expect(writtenEvent).toMatchObject({
       stream_id: mockPurpose.id,
       version: "1",
-      type: "PurposeArchived",
+      type: "WaitingForApprovalPurposeDeleted",
       event_version: 2,
     });
 
-    const expectedPurpose: Purpose = {
-      ...mockPurpose,
-      versions: [
-        {
-          ...mockPurposeVersion1,
-          state: purposeVersionState.archived,
-          updatedAt: new Date(),
-        },
-      ],
-      updatedAt: new Date(),
-    };
-
     const writtenPayload = decodeProtobufPayload({
-      messageType: PurposeArchivedV2,
+      messageType: WaitingForApprovalPurposeDeletedV2,
       payload: writtenEvent.data,
     });
 
-    expect(writtenPayload.purpose).toEqual(toPurposeV2(expectedPurpose));
-    expect(
-      writtenPayload.purpose?.versions.find(
-        (v) => v.id === returnedPurposeVersion.id
-      )
-    ).toEqual(toPurposeVersionV2(returnedPurposeVersion));
-
-    vi.useRealTimers();
+    expect(writtenPayload.purpose).toEqual(toPurposeV2(mockPurpose));
   });
-  it("should write on event-store for the archiving of a purpose version when requester is Consumer Delegate", async () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date());
-
-    const mockPurposeVersion: PurposeVersion = {
-      ...getMockPurposeVersion(),
-      state: purposeVersionState.active,
-    };
+  it("should succeed when requester is Consumer Delegate and the Purpose is in a deletable state", async () => {
+    const authData = getMockAuthData();
+    const mockEService = getMockEService();
+    const mockPurposeVersion = getMockPurposeVersion(purposeVersionState.draft);
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
+      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
       delegationId: generateId<DelegationId>(),
     };
-
-    const authData = getMockAuthData();
 
     const delegation = getMockDelegation({
       id: mockPurpose.delegationId,
@@ -197,12 +173,10 @@ describe("archivePurposeVersion", () => {
     await addOnePurpose(mockPurpose);
     await addOneDelegation(delegation);
     await addSomeRandomDelegations(mockPurpose, addOneDelegation);
+    await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
-    const returnedPurposeVersion = await purposeService.archivePurposeVersion(
-      {
-        purposeId: mockPurpose.id,
-        versionId: mockPurposeVersion.id,
-      },
+    await purposeService.deletePurpose(
+      mockPurpose.id,
       getMockContext({ authData })
     );
 
@@ -211,40 +185,18 @@ describe("archivePurposeVersion", () => {
     expect(writtenEvent).toMatchObject({
       stream_id: mockPurpose.id,
       version: "1",
-      type: "PurposeArchived",
+      type: "DraftPurposeDeleted",
       event_version: 2,
     });
 
-    const expectedPurpose: Purpose = {
-      ...mockPurpose,
-      versions: [
-        {
-          ...mockPurposeVersion,
-          state: purposeVersionState.archived,
-          updatedAt: new Date(),
-        },
-      ],
-      updatedAt: new Date(),
-    };
-
     const writtenPayload = decodeProtobufPayload({
-      messageType: PurposeArchivedV2,
+      messageType: DraftPurposeDeletedV2,
       payload: writtenEvent.data,
     });
 
-    expect(writtenPayload.purpose).toEqual(toPurposeV2(expectedPurpose));
-    expect(
-      writtenPayload.purpose?.versions.find(
-        (v) => v.id === returnedPurposeVersion.id
-      )
-    ).toEqual(toPurposeVersionV2(returnedPurposeVersion));
-
-    vi.useRealTimers();
+    expect(writtenPayload.purpose).toEqual(toPurposeV2(mockPurpose));
   });
-  it("should write on event-store for the archiving of a purpose version when requester is Consumer Delegate and the eservice was created by a delegated tenant", async () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date());
-
+  it("should succeed when requester is Consumer Delegate and the eservice was created by a delegated tenant and the Purpose is in a deletable state", async () => {
     const producer = {
       ...getMockTenant(),
       id: generateId<TenantId>(),
@@ -268,7 +220,6 @@ describe("archivePurposeVersion", () => {
 
     const eservice: EService = {
       ...getMockEService(),
-      mode: eserviceMode.receive,
       producerId: producer.id,
     };
     const agreement: Agreement = {
@@ -276,13 +227,9 @@ describe("archivePurposeVersion", () => {
       producerId: producer.id,
       consumerId: consumer.id,
       eserviceId: eservice.id,
-      state: agreementState.active,
     };
 
-    const mockPurposeVersion: PurposeVersion = {
-      ...getMockPurposeVersion(),
-      state: purposeVersionState.active,
-    };
+    const mockPurposeVersion = getMockPurposeVersion(purposeVersionState.draft);
 
     const delegatePurpose: Purpose = {
       ...getMockPurpose(),
@@ -320,11 +267,8 @@ describe("archivePurposeVersion", () => {
     await addOneDelegation(consumerDelegation);
     await addSomeRandomDelegations(delegatePurpose, addOneDelegation);
 
-    const returnedPurposeVersion = await purposeService.archivePurposeVersion(
-      {
-        purposeId: delegatePurpose.id,
-        versionId: mockPurposeVersion.id,
-      },
+    await purposeService.deletePurpose(
+      delegatePurpose.id,
       getMockContext({ authData: getMockAuthData(consumerDelegate.id) })
     );
 
@@ -333,139 +277,92 @@ describe("archivePurposeVersion", () => {
     expect(writtenEvent).toMatchObject({
       stream_id: delegatePurpose.id,
       version: "1",
-      type: "PurposeArchived",
+      type: "DraftPurposeDeleted",
       event_version: 2,
     });
 
-    const expectedPurpose: Purpose = {
-      ...delegatePurpose,
-      versions: [
-        {
-          ...mockPurposeVersion,
-          state: purposeVersionState.archived,
-          updatedAt: new Date(),
-        },
-      ],
-      updatedAt: new Date(),
-    };
-
     const writtenPayload = decodeProtobufPayload({
-      messageType: PurposeArchivedV2,
+      messageType: DraftPurposeDeletedV2,
       payload: writtenEvent.data,
     });
 
-    expect(writtenPayload.purpose).toEqual(toPurposeV2(expectedPurpose));
-    expect(
-      writtenPayload.purpose?.versions.find(
-        (v) => v.id === returnedPurposeVersion.id
-      )
-    ).toEqual(toPurposeVersionV2(returnedPurposeVersion));
-
-    vi.useRealTimers();
+    expect(writtenPayload.purpose).toEqual(toPurposeV2(delegatePurpose));
   });
-
   it("should throw purposeNotFound if the purpose doesn't exist", async () => {
-    const randomPurposeId: PurposeId = generateId();
-    const randomVersionId: PurposeVersionId = generateId();
+    const randomId: PurposeId = generateId();
     const mockPurpose = getMockPurpose();
-    await addOnePurpose(mockPurpose);
 
+    await addOnePurpose(mockPurpose);
     expect(
-      purposeService.archivePurposeVersion(
-        {
-          purposeId: randomPurposeId,
-          versionId: randomVersionId,
-        },
+      purposeService.deletePurpose(
+        randomId,
         getMockContext({ authData: getMockAuthData(mockPurpose.consumerId) })
       )
-    ).rejects.toThrowError(purposeNotFound(randomPurposeId));
+    ).rejects.toThrowError(purposeNotFound(randomId));
   });
   it("should throw organizationIsNotTheConsumer if the requester is not the consumer", async () => {
-    const randomOrganizationId: TenantId = generateId();
-
-    const mockPurposeVersion: PurposeVersion = {
-      ...getMockPurposeVersion(),
-      state: purposeVersionState.active,
-    };
+    const mockEService = getMockEService();
+    const mockPurposeVersion: PurposeVersion = getMockPurposeVersion(
+      purposeVersionState.draft
+    );
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
+      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
     };
 
     await addOnePurpose(mockPurpose);
+    await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
     expect(
-      purposeService.archivePurposeVersion(
-        {
-          purposeId: mockPurpose.id,
-          versionId: mockPurposeVersion.id,
-        },
-        getMockContext({ authData: getMockAuthData(randomOrganizationId) })
-      )
-    ).rejects.toThrowError(organizationIsNotTheConsumer(randomOrganizationId));
-  });
-  it("should throw purposeVersionNotFound if the purpose version doesn't exist", async () => {
-    const randomVersionId: PurposeVersionId = generateId();
-    const mockPurpose: Purpose = {
-      ...getMockPurpose(),
-      versions: [],
-    };
-
-    await addOnePurpose(mockPurpose);
-
-    expect(
-      purposeService.archivePurposeVersion(
-        {
-          purposeId: mockPurpose.id,
-          versionId: randomVersionId,
-        },
-        getMockContext({ authData: getMockAuthData(mockPurpose.consumerId) })
+      purposeService.deletePurpose(
+        mockPurpose.id,
+        getMockContext({ authData: getMockAuthData(mockEService.producerId) })
       )
     ).rejects.toThrowError(
-      purposeVersionNotFound(mockPurpose.id, randomVersionId)
+      organizationIsNotTheConsumer(mockEService.producerId)
     );
   });
   it.each(
     Object.values(purposeVersionState).filter(
       (state) =>
-        state !== purposeVersionState.active &&
-        state !== purposeVersionState.suspended
+        state !== purposeVersionState.waitingForApproval &&
+        state !== purposeVersionState.draft
     )
   )(
-    "should throw notValidVersionState if the purpose version is in %s state",
+    "should throw purposeCannotBeDeleted if the purpose has a $s version ",
     async (state) => {
+      const mockEService = getMockEService();
       const mockPurposeVersion = getMockPurposeVersion(state);
 
       const mockPurpose: Purpose = {
         ...getMockPurpose(),
+        eserviceId: mockEService.id,
         versions: [mockPurposeVersion],
       };
 
       await addOnePurpose(mockPurpose);
+      await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
       expect(
-        purposeService.archivePurposeVersion(
-          {
-            purposeId: mockPurpose.id,
-            versionId: mockPurposeVersion.id,
-          },
+        purposeService.deletePurpose(
+          mockPurpose.id,
           getMockContext({
             authData: getMockAuthData(mockPurpose.consumerId),
           })
         )
-      ).rejects.toThrowError(
-        notValidVersionState(mockPurposeVersion.id, mockPurposeVersion.state)
-      );
+      ).rejects.toThrowError(purposeCannotBeDeleted(mockPurpose.id));
     }
   );
-  it("should throw organizationIsNotTheDelegatedConsumer when the requester is the Consumer and is archiving a purpose version created by the delegate in archivePurposeVersion", async () => {
+  it("should throw organizationIsNotTheDelegatedConsumer when the requester is the Consumer and is deleting a purpose created by the delegate in deletePurpose", async () => {
     const authData = getMockAuthData();
-    const mockPurposeVersion: PurposeVersion = {
-      ...getMockPurposeVersion(),
-      state: purposeVersionState.active,
-    };
+    const mockEService = getMockEService();
+    const mockPurposeVersion: PurposeVersion = getMockPurposeVersion(
+      purposeVersionState.draft
+    );
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
+      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
       consumerId: authData.organizationId,
       delegationId: generateId<DelegationId>(),
@@ -481,23 +378,19 @@ describe("archivePurposeVersion", () => {
     });
     await addOnePurpose(mockPurpose);
     await addOneDelegation(delegation);
+    await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
     expect(
-      purposeService.archivePurposeVersion(
-        {
-          purposeId: mockPurpose.id,
-          versionId: mockPurposeVersion.id,
-        },
-        getMockContext({ authData })
-      )
+      purposeService.deletePurpose(mockPurpose.id, getMockContext({ authData }))
     ).rejects.toThrowError(
       organizationIsNotTheDelegatedConsumer(
         authData.organizationId,
-        mockPurpose.delegationId
+        delegation.id
       )
     );
   });
-  it("should throw puroposeDelegationNotFound when the requester is the Consumer, is archiving a purpose created by a delegate, but the delegation cannot be found", async () => {
+
+  it("should throw puroposeDelegationNotFound when the requester is the Consumer, is deleting a purpose created by a delegate in deletePurpose, but the delegation cannot be found", async () => {
     const authData = getMockAuthData();
     const mockEService = getMockEService();
     const mockPurposeVersion: PurposeVersion = getMockPurposeVersion(
@@ -512,31 +405,27 @@ describe("archivePurposeVersion", () => {
     };
 
     await addOnePurpose(mockPurpose);
-    await addOneEService(mockEService);
+    await writeInReadmodel(toReadModelEService(mockEService), eservices);
+
     expect(
-      purposeService.archivePurposeVersion(
-        {
-          purposeId: mockPurpose.id,
-          versionId: mockPurposeVersion.id,
-        },
-        getMockContext({ authData })
+      purposeService.deletePurpose(
+        mockPurpose.id,
+        getMockContext({ authData: getMockAuthData(mockEService.producerId) })
       )
     ).rejects.toThrowError(
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       puroposeDelegationNotFound(mockPurpose.id, mockPurpose.delegationId!)
     );
   });
-
-  it("should throw organizationIsNotTheConsumer when the requester is a delegate for the eservice and there is no delegationId in the purpose in archivePurposeVersion", async () => {
-    const authData = getMockAuthData();
-    const mockPurposeVersion: PurposeVersion = {
-      ...getMockPurposeVersion(),
-      state: purposeVersionState.active,
-    };
+  it("should throw organizationIsNotTheConsumer if the requester is a delegate for the eservice and there is no delegationId in the purpose", async () => {
+    const mockEService = getMockEService();
+    const mockPurposeVersion: PurposeVersion = getMockPurposeVersion(
+      purposeVersionState.draft
+    );
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
+      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
-      consumerId: authData.organizationId,
       delegationId: undefined,
     };
 
@@ -549,29 +438,25 @@ describe("archivePurposeVersion", () => {
     });
     await addOnePurpose(mockPurpose);
     await addOneDelegation(delegation);
+    await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
     expect(
-      purposeService.archivePurposeVersion(
-        {
-          purposeId: mockPurpose.id,
-          versionId: mockPurposeVersion.id,
-        },
+      purposeService.deletePurpose(
+        mockPurpose.id,
         getMockContext({ authData: getMockAuthData(delegation.delegateId) })
       )
     ).rejects.toThrowError(organizationIsNotTheConsumer(delegation.delegateId));
   });
 
   it("should throw organizationIsNotTheDelegatedConsumer if the the requester is a delegate for the eservice and there is a delegationId in purpose but for a different delegationId (a different delegate)", async () => {
-    const authData = getMockAuthData();
     const mockEService = getMockEService();
-    const mockPurposeVersion: PurposeVersion = {
-      ...getMockPurposeVersion(),
-      state: purposeVersionState.active,
-    };
+    const mockPurposeVersion: PurposeVersion = getMockPurposeVersion(
+      purposeVersionState.draft
+    );
     const mockPurpose: Purpose = {
       ...getMockPurpose(),
+      eserviceId: mockEService.id,
       versions: [mockPurposeVersion],
-      consumerId: authData.organizationId,
       delegationId: generateId<DelegationId>(),
     };
 
@@ -596,14 +481,11 @@ describe("archivePurposeVersion", () => {
     await addOnePurpose(mockPurpose);
     await addOneDelegation(delegation);
     await addOneDelegation(purposeDelegation);
-    await addOneEService(mockEService);
+    await writeInReadmodel(toReadModelEService(mockEService), eservices);
 
     expect(
-      purposeService.archivePurposeVersion(
-        {
-          purposeId: mockPurpose.id,
-          versionId: mockPurposeVersion.id,
-        },
+      purposeService.deletePurpose(
+        mockPurpose.id,
         getMockContext({ authData: getMockAuthData(delegation.delegateId) })
       )
     ).rejects.toThrowError(
