+++ packages/purpose-process/vitest.integration.config.ts
@@ -2,7 +2,8 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    globalSetup: ["./test/vitestGlobalSetup.ts"],
+    globalSetup: ["./test/vitestIntegrationGlobalSetup.ts"],
+    include: ["./test/integration/**/*.test.ts"],
     testTimeout: 60000,
     hookTimeout: 60000,
     fileParallelism: false,
